Merges cherry-pick-from-phyta (origin/auto-debugger/work + AUT-553/AUT-527/SEN0220 upstream)
into feature/phyta-frontend. Conflicts in Phyta Frontend/ resolved with ours strategy
(feature/phyta-frontend has newer orbital+grid+wizard state from 240ca9a3).

+++ El Trabajante/src/utils/time_manager.cpp
@@ -359,7 +359,7 @@ void TimeManager::refreshPrimaryNtpFromGateway() {
         return;
     }
     const IPAddress gw = WiFi.gatewayIP();
-    if (gw == IPAddress(0u)) {
+    if (gw == IPAddress(static_cast<uint32_t>(0))) {
         ntp_server_primary_ = NTP_SERVER_PRIMARY;
         return;
     }
